+++ components/src/main/java/uk/gov/hmrc/components/base/PaddedComponent.kt
@@ -16,5 +16,6 @@
 package uk.gov.hmrc.components.base
 
 interface PaddedComponent {
+    fun adjustPaddingForChildren() {}
     fun removeChildPadding()
 }